Cleanup

+++ app/src/main/java/io/esalenko/wirextest/db/RemoteKeysEntity.kt
@@ -6,5 +6,5 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "remote_keys")
 data class RemoteKeysEntity(
     @PrimaryKey(autoGenerate = false)
-    val id: Long, val nextKey: Int
+    val id: Int, val nextKey: Int
 )
